package plural renamed to singular
