+++ MovieGPT-api/src/main/java/lt/techin/group/project/exception/GenreNotFoundException.java
@@ -1,11 +1,11 @@
-package lt.techin.group.project.exceptions;
+package lt.techin.group.project.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
-public class UserNotFoundException extends RuntimeException {
-  public UserNotFoundException(String message) {
+public class GenreNotFoundException extends RuntimeException {
+  public GenreNotFoundException(String message) {
     super(message);
   }
 }
